% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coeff.R
\name{coeff}
\alias{coeff}
\title{Function to estimate coefficients at time t}
\usage{
coeff(j, T1, T2, x, y)
}
\arguments{
\item{j}{a number indicating time point of observation}

\item{T1}{a vector indicating assignment to treatment 1}

\item{T2}{a vector indicating assignment to treatment 2}

\item{x}{matrix of mediator values in wide format}

\item{y}{matrix of outcome outcomes in wide format}
}
\value{
\item{coeff.est}{estimated coefficients of the mediation model}

\item{nomissing.index}{index of complete cases}
}
\description{
Part of the set of internal functions called within the \code{tvmcurve_3trt} function 
to assist in the estimation of time varying mediation effect.
}
