% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estCoeff.R
\name{estCoeff}
\alias{estCoeff}
\title{Function to estimate coefficients at time t.}
\usage{
estCoeff(newMO.j.est)
}
\arguments{
\item{newMO.j.est}{a list containing mean centered mediators and outcomes}
}
\value{
\item{coeff.est}{estimated coefficients of the mediation model}
}
\description{
Part of the set of internal functions called within the \code{tvma} function to assist 
in the estimation of the time varying mediation effect.
}
