\name{estErrorCovar}
\alias{estErrorCovar}
\title{Internal tvmediation function}

\description{
Internal tvmediation function called by other functions. Used to estimate error covariances
} 

\usage{
est.error.covar <- estErrorCovar()
}


\details{
This function is not desgined to be called directly by the user.
}

\seealso{
\code{\link{tvmediation}}, \code{\link{tvma}}, \code{\link{LongToWide}}, \code{\link{smoker}}
}

\keyword{internal}