\name{kern}
\alias{kern}
\title{Internal tvmediation function}

\description{
Internal tvmediation function called by other functions.
} 

\usage{
k <- kern()
}


\details{
This function is not desgined to be called directly by the user.
}

\seealso{
\code{\link{tvmediation}}, \code{\link{tvma}}, \code{\link{LongToWide}}, \code{\link{smoker}}
}

\keyword{internal}