\name{localLinearWt}
\alias{localLinearWt}
\title{Internal tvmediation function}

\description{
Internal tvmediation function called by other functions. Returns weights of local polynomial regression
} 

\usage{
weight <- localLinearWt()
}


\details{
This function is not desgined to be called directly by the user.
}

\seealso{
\code{\link{tvmediation}}, \code{\link{tvma}}, \code{\link{LongToWide}}, \code{\link{smoker}}
}

\keyword{internal}