\name{tvmediation}
\title{Time Varying Mediation Analysis}
\alias{tvmediation}

\description{

Provides functions for estimating mediation effects that vary over 
time as described in Cai, X., Piper, M. E., Li, R., & Coffman, D. L. (2020). 
Estimation and inference for the mediation effect in a time-varying mediation 
model. \url{https://arxiv.org/abs/2008.11797}
} 

\seealso{
\code{\link{tvma}}, \code{\link{tvma_3trt}}, \code{\link{tvmb}}, \code{\link{LongToWide}}

Useful links:
\itemize{
  \item \url{https://github.com/dcoffman/tvmediation/wiki}
  \item \url{https://github.com/dcoffman/tvmediation/issues}
}
}

\author{
\strong{Maintainer}: Donna L. Coffman \email{dcoffman@temple.edu}

Authors:
	\itemize{
	  \item Donna L. Coffman
		\item Xizhen Cai
		\item Yajnaseni Chakraborti
		\item Harry Zobel
	}
}
\keyword{internal}