% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootci_coeff_2trt.R
\name{bootci_coeff_2trt}
\alias{bootci_coeff_2trt}
\title{Bootstrapping samples to estimate confidence intervals for coefficients of continuous outcome and two treatment (exposure) groups}
\usage{
bootci_coeff_2trt(trt, t.seq, M, Y, t.est, deltat, replicates)
}
\arguments{
\item{trt}{a vector with treatment values}

\item{t.seq}{a vector of time points for each observation}

\item{M}{matrix of mediator values in wide format}

\item{Y}{matrix of outcome outcomes in wide format}

\item{t.est}{time points to make the estimation                              Default = t.seq}

\item{deltat}{a small constant which controls the time-lag of the effect of the mediator on the outcome.}

\item{replicates}{number of replicates for bootstrapping confidence intervals.    Default = 1000}
}
\value{
\item{CI.upper.alpha}{upper limit of confidence intervals for coefficient hat.alpha}

\item{CI.lower.alpha}{lower limit of confidence intervals for coefficient hat.alpha}

\item{CI.upper.gamma}{upper limit of confidence intervals for coefficient hat.gamma}

\item{CI.lower.gamma}{lower limit of confidence intervals for coefficient hat.gamma}

\item{CI.upper.beta}{upper limit of confidence intervals for coefficient hat.beta}

\item{CI.lower.beta}{lower limit of confidence intervals for coefficient hat.beta}
}
\description{
Part of the set of internal functions for estimating bootstrapped confidence intervals for the coefficients of the mediation model for continuous outcome and two treatment (exposure) groups.
}
