% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootci_coeff_3trt.R
\name{bootci_coeff_3trt}
\alias{bootci_coeff_3trt}
\title{Bootstrapping samples to estimate confidence intervals for coefficients of continuous outcome and three treatment (exposure) groups}
\usage{
bootci_coeff_3trt(
  NRT1,
  NRT2,
  t.seq,
  mediator,
  outcome,
  t.est,
  original.coeff,
  boot.sample = 1000
)
}
\arguments{
\item{NRT1}{a vector with treatment arm 1 values}

\item{NRT2}{a vector with treatment arm 2 values}

\item{t.seq}{a vector of time points for each observation}

\item{mediator}{matrix of mediator values in wide format}

\item{outcome}{matrix of outcome outcomes in wide format}

\item{t.est}{time points to make the estimation                              Default = t.seq.}

\item{original.coeff}{a list of the estimated coefficients \code{hat.alpha1}, \code{hat.alpha2} and \code{hat.beta3}.}

\item{boot.sample}{number of replicates for bootstrapping confidence intervals.    Default = 1000.}
}
\value{
\item{alw1}{lower limit of confidence intervals for estimated NRT1 effect on mediator}

\item{aup1}{upper limit of confidence intervals for estimated NRT1 effect on mediator}

\item{alw2}{lower limit of confidence intervals for estimated NRT2 effect on mediator}

\item{aup2}{upper limit of confidence intervals for estimated NRT2 effect on mediator}

\item{blw3}{lower limit of confidence intervals for estimated effect of mediator on outcome}

\item{bup3}{upper limit of confidence intervals for estimated effect of mediator on outcome}
}
\description{
Part of the set of internal functions for estimating bootstrapped confidence intervals for the coefficients of the mediation model for continuous outcome and three treatment (exposure) groups.
}
