% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootci_coeff_binary.R
\name{bootci_coeff_binary}
\alias{bootci_coeff_binary}
\title{Bootstrapping samples to estimate confidence intervals for binary outcome coefficients}
\usage{
bootci_coeff_binary(treatment, t.seq, m, outcome, replicates = 1000)
}
\arguments{
\item{treatment}{a vector with treatment values}

\item{t.seq}{a vector of unique time points for each observation}

\item{m}{matrix of mediator values in wide format}

\item{outcome}{matrix of outcome outcomes in wide format}

\item{replicates}{Number of replicates for bootstrapping confidence intervals.    Default = 1000.}
}
\value{
\item{t.seq}{time points of estimation}

\item{CI.lower.a1}{lower limit of confidence intervals for alpha1_hat}

\item{CI.upper.a1}{upper limit of confidence intervals for alpha1_hat}

\item{CI.lower.b1}{lower limit of confidence intervals for beta1_hat}

\item{CI.upper.b1}{upper limit of confidence intervals for beta1_hat}

\item{CI.lower.b2}{lower limit of confidence intervals for beta2_hat}

\item{CI.upper.b2}{upper limit of confidence intervals for beta2_hat}
}
\description{
Internal function for estimating bootstrapped confidence intervals for the coefficients of the mediation model for binary outcome.
}
