% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootci_tvmb.R
\name{bootci_tvmb}
\alias{bootci_tvmb}
\title{Bootstrapping samples to estimate confidence intervals for binary outcome mediation effects}
\usage{
bootci_tvmb(treatment, t.seq, m, outcome, coeff_data, replicates = 1000)
}
\arguments{
\item{treatment}{a vector with treatment values}

\item{t.seq}{a vector of unique time points for each observation}

\item{m}{matrix of mediator values in wide format}

\item{outcome}{matrix of outcome outcomes in wide format}

\item{coeff_data}{a merged dataset of indirect and direct effects coefficients and confidence intervals estimated from \code{bootci_coeff_binary}}

\item{replicates}{number of replicates for bootstrapping confidence intervals.    Default = 1000.}
}
\value{
\item{timeseq}{time points of estimation}

\item{alpha_hat}{estimated exposure effect on mediator (indirect effect)}

\item{CI.lower.a}{lower limit of confidence intervals for estimated coefficient alpha_hat}

\item{CI.upper.a}{upper limit of confidence intervals for estimated coefficient alpha_hat}

\item{gamma_hat}{estimated exposure effect on outcome (direct effect)}

\item{CI.lower.g}{lower limit of confidence intervals for estimated coefficient gamma_hat}

\item{CI.upper.g}{upper limit of confidence intervals for estimated coefficient gamma_hat}

\item{beta_hat}{estimated mediation effect on outcome (indirect effect)}

\item{CI.lower.b}{lower limit of confidence intervals for estimated coefficient beta_hat}

\item{CI.upper.b}{upper limit of confidence intervals for estimated coefficient beta_hat}

\item{tao_hat}{estimated exposure effect on outcome, excluding adjustment for mediator (total effect)}

\item{CI.lower.t}{lower limit of confidence intervals for estimated coefficient tao_hat}

\item{CI.upper.t}{upper limit of confidence intervals for estimated coefficient tao_hat}

\item{medEffect}{time varying mediation effect (product term)}

\item{CI.lower}{lower limit of confidence intervals for medEffect}

\item{CI.upper}{upper limit of confidence intervals for medEffect}
}
\description{
Part of the set of internal functions for estimating bootstrapped confidence intervals for the mediation effect estimate for binary outcome when user argument \code{CI="boot"}.
}
