% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootci_tvmb.R
\name{bootci_tvmb}
\alias{bootci_tvmb}
\title{Bootstrapping samples to estimate confidence intervals for binary outcome mediation effects}
\usage{
bootci_tvmb(treatment, t.seq, m, outcome, coeff_data, replicates = 1000)
}
\arguments{
\item{treatment}{a vector with treatment values}

\item{t.seq}{a vector of unique time points for each observation}

\item{m}{matrix of mediator values in wide format}

\item{outcome}{matrix of outcome outcomes in wide format}

\item{coeff_data}{a merged dataset of indirect and direct effects coefficients and confidence intervals estimated from \code{bootci_coeff_binary}}

\item{replicates}{number of replicates for bootstrapping confidence intervals.    Default = 1000.}
}
\value{
\item{timeseq}{time points of estimation}

\item{alpha1_hat}{exposure effect on mediator (indirect effect)}

\item{CI.lower.a1}{lower limit of confidence intervals for alpha1_hat}

\item{CI.upper.a1}{upper limit of confidence intervals for alpha1_hat}

\item{beta2_hat}{estimated mediation effect on outcome (indirect effect)}

\item{CI.lower.b2}{lower limit of confidence intervals for beta2_hat}

\item{CI.upper.b2}{upper limit of confidence intervals for beta2_hat}

\item{b1All}{estimated exposure effect on outcome (direct effect)}

\item{cAll}{estimated effect of exposure on outcome (total effect)}

\item{medDiff}{time varying mediation effect (difference term)}

\item{medEffect}{time varying mediation effect (product term)}

\item{CI.low}{lower limit of confidence intervals for medEffect}

\item{CI.upper}{upper limit of confidence intervals for medEffect}
}
\description{
Part of the set of internal functions for estimating bootstrapped confidence intervals for the mediation effect estimate for binary outcome when user argument \code{CI="boot"}.
}
