% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coeff.R
\name{coeff}
\alias{coeff}
\title{Function to estimate coefficients at time T}
\usage{
coeff(j, NRT1, NRT2, x, y)
}
\arguments{
\item{j}{a number indicating time point of observation}

\item{NRT1}{a vector with treatment arm 1 values}

\item{NRT2}{a vector with treatment arm 2 values}

\item{x}{matrix of mediator values in wide format}

\item{y}{matrix of outcome outcomes in wide format}
}
\value{
\item{coeff.est}{estimated coefficients of the mediation model}

\item{nomissing.index}{index of complete cases}
}
\description{
Part of the set of internal functions called within the \code{tvmcurve_3trt} function to assist in the estimation of time varying mediation effect.
}
