% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estBootCIs.R
\name{estBootCIs}
\alias{estBootCIs}
\title{Bootstrapping samples to estimate confidence intervals for continuous outcome and two treatment (exposure) groups mediation effects}
\usage{
estBootCIs(trt, t.seq, M, Y, t.est, deltat, replicates)
}
\arguments{
\item{trt}{a vector with treatment values}

\item{t.seq}{a vector of time points at each obs}

\item{M}{matrix of mediator values}

\item{Y}{matrix of outcome outcomes}

\item{t.est}{time points to make the estimation}

\item{deltat}{a small constant which controls the time-lag of the effect of the mediator on the outcome.}

\item{replicates}{number of replicates for bootstrapping confidence intervals.}
}
\value{
{boot.se}{bootstrapped standard error for the estimated mediation effect}

{CI.upper}{upper limit of confidence intervals for the estimated mediation effect}

{CI.lower}{lower limit of confidence intervals for the estimated mediation effect}
}
\description{
Part of the set of internal functions for estimating bootstrapped confidence intervals for the mediation model of continuous outcome and two treatment (exposure) groups when user argument \code{CI="boot"}.
}
