\name{estBootCIs}
\alias{estBootCIs}
\title{Internal tvmediation function}

\description{
Internal tvmediation function called by other functions. Used to derive bootstrap confidence intervals
} 

\usage{
est.boot <- estBootCIs()
}


\details{
This function is not desgined to be called directly by the user.
}

\seealso{
\code{\link{tvmediation}}, \code{\link{tvma}}, \code{\link{LongToWide}}, \code{\link{smoker}}
}

\keyword{internal}