% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newMediatorOutcome.R
\name{newMediatorOutcome}
\alias{newMediatorOutcome}
\title{Function to compute new Mediator and Outcome using time T and T-1 mean centered on the individual}
\usage{
newMediatorOutcome(trt, M, Y)
}
\arguments{
\item{trt}{numeric binary treatment schedule for each individual}

\item{mediator}{(t.seq x N) matrix where N = number of observations. Column 1 is mediator at time T-1. Column 2 is mediator at time T.}

\item{outcome}{(Nx1) matrix were N = number of observations. Column 1 is outcome at time T-1.}

\item{centerOutcome}{outcomes must be centered estimating coefficients and covariance of error terms.}
}
\value{
\item{newMO}{list containing new mediators, outcomes, and index of complete cases}
}
\description{
Part of the set of internal functions called within the \code{tvma} function to assist in the estimation of time varying mediation effect.
}
