\name{newMediatorOutcome}
\alias{newMediatorOutcome}
\title{Internal tvmediation function}

\description{
Internal tvmediation function called by other functions. Derives mean-centered mediators and outcomes
} 

\usage{
newMO <- newMediatorOutcome()
}


\details{
This function is not desgined to be called directly by the user.
}

\seealso{
\code{\link{tvmediation}}, \code{\link{tvma}}, \code{\link{LongToWide}}, \code{\link{smoker}}
}

\keyword{internal}