\name{smoothest}
\alias{smoothest}
\title{Internal tvmediation function}

\description{
Internal tvmediation function called by other functions. Computes local polynomial estimation using rule of thumb for bandwidth selection
} 

\usage{
results <- smoothest()
}


\details{
This function is not desgined to be called directly by the user.
}

\seealso{
\code{\link{tvmediation}}, \code{\link{tvma}}, \code{\link{LongToWide}}, \code{\link{smoker}}
}

\keyword{internal}