% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tvma_3trt.R
\name{tvma_3trt}
\alias{tvma_3trt}
\title{Time-varying mediation function for continuous outcome and 3 treatment arms (exposure groups)}
\usage{
tvma_3trt(
  NRT1,
  NRT2,
  t.seq,
  mediator,
  outcome,
  t.est = t.seq,
  plot = FALSE,
  CI = "boot",
  replicates = 1000,
  verbose = FALSE
)
}
\description{
Time-varying mediation function for continuous outcome and 3 treatment arms (exposure groups)
}
