% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tvma_3trt.R
\name{tvma_3trt}
\alias{tvma_3trt}
\title{Time Varying Mediation Function: Continuous Outcome and Three Treatment Arms (Exposure Groups)}
\usage{
tvma_3trt(
  NRT1,
  NRT2,
  t.seq,
  mediator,
  outcome,
  t.est = t.seq,
  plot = FALSE,
  CI = "boot",
  replicates = 1000,
  grpname = "NRT",
  verbose = FALSE
)
}
\arguments{
\item{NRT1}{a vector with treatment arm 1 (exposure group 1) values}

\item{NRT2}{a vector with treatment arm 2 (exposure group 2) values}

\item{t.seq}{a vector of time points for each observation}

\item{mediator}{matrix of mediator values in wide format}

\item{outcome}{matrix of outcome outcomes in wide format}

\item{t.est}{a vector of time points to make the estimation                                  Default = t.seq. (OPTIONAL ARGUMENT)}

\item{plot}{TRUE or FALSE for plotting mediation effect                                     Default = "FALSE". (OPTIONAL ARGUMENT)}

\item{CI}{"none" or "boot" method of deriving confidence intervals.                       Default = "boot". (OPTIONAL ARGUMENT)}

\item{replicates}{number of replicates for bootstrapping confidence intervals.                    Default = 1000. (OPTIONAL ARGUMENT)}

\item{grpname}{name of the treatment arms (exposure groups) to be displayed in the results.    Default = "NRT". (OPTIONAL ARGUMENT)}

\item{verbose}{TRUE or FALSE for printing results to screen.                                   Default = "FALSE". (OPTIONAL ARGUMENT)}
}
\value{
\item{hat.alpha1}{estimated effect of treatment arm 1 (exposure group 1) on mediator (indirect effect component)}

\item{CI.lower.alpha1}{lower limit of confidence intervals for estimated coefficient hat.alpha1}

\item{CI.upper.alpha1}{upper limit of confidence intervals for estimated coefficient hat.alpha1}

\item{hat.alpha2}{estimated effect of treatment arm 2 (exposure group 2) on mediator (indirect effect component)}

\item{CI.lower.alpha2}{lower limit of confidence intervals for estimated coefficient hat.alpha2}

\item{CI.upper.alpha2}{upper limit of confidence intervals for estimated coefficient hat.alpha2}

\item{hat.gamma1}{estimated effect of treatment arm 1 (exposure group 1) on outcome (direct effect)}

\item{CI.lower.gamma1}{lower limit of confidence intervals for estimated coefficient hat.gamma1}

\item{CI.upper.gamma1}{upper limit of confidence intervals for estimated coefficient hat.gamma1}

\item{hat.gamma2}{estimated effect of treatment arm 2 (exposure group 2) on outcome (direct effect)}

\item{CI.lower.gamma2}{lower limit of confidence intervals for estimated coefficient hat.gamma2}

\item{CI.upper.gamma2}{upper limit of confidence intervals for estimated coefficient hat.gamma2}

\item{hat.tao1}{estimated effect of treatment arm 1 (exposure group 1) on outcome, excluding adjustment for mediator (total effect)}

\item{CI.lower.tao1}{lower limit of confidence intervals for estimated coefficient hat.tao1}

\item{CI.upper.tao1}{upper limit of confidence intervals for estimated coefficient hat.tao1}

\item{hat.tao2}{estimated effect of treatment arm 2 (exposure group 2) on outcome, excluding adjustment for mediator (total effect)}

\item{CI.lower.tao2}{lower limit of confidence intervals for estimated coefficient hat.tao2}

\item{CI.upper.tao2}{upper limit of confidence intervals for estimated coefficient hat.tao2}

\item{hat.beta}{estimated mediator effect on outcome (indirect effect component)}

\item{CI.lower.beta}{lower limit of confidence intervals for estimated coefficient hat.beta}

\item{CI.upper.beta}{upper limit of confidence intervals for estimated coefficient hat.beta}

\item{hat.mediation1}{time varying mediation effect - treatment arm 1 (exposure group 1) on outcome}

\item{SE_MedEff1}{estimated standard errors of hat.mediation1}

\item{CI.upper.NRT1}{upper limit of confidence intervals for hat.mediation1}

\item{CI.lower.NRT1}{lower limit of confidence intervals for hat.mediation1}

\item{hat.mediation2}{time varying mediation effect - treatment arm 2 (exposure group 2) on outcome}

\item{SE_MedEff2}{estimated standard errors of hat.mediation2}

\item{CI.upper.NRT2}{upper limit of confidence intervals for hat.mediation2}

\item{CI.lower.NRT2}{lower limit of confidence intervals for hat.mediation2}
}
\description{
Function to estimate the time-varying mediation effect and bootstrap standard errors, involving three treatment/exposure groups and continuous outcome.
}
\section{Plot Returns}{

\enumerate{
\item{\code{plot1_a1 }}{plot for hat.alpha1 across t.est with CIs}
\item{\code{plot2_a2 }}{plot for hat.alpha2 across t.est with CIs}
\item{\code{plot3_g1 }}{plot for hat.gamma1 across t.est with CIs}
\item{\code{plot4_g2 }}{plot for hat.gamma2 across t.est with CIs}
\item{\code{plot5_t1 }}{plot for hat.tao1 across t.est with CIs}
\item{\code{plot6_t2 }}{plot for hat.tao2 across t.est with CIs}
\item{\code{plot7_b }}{plot for hat.beta across t.est with CIs}
\item{\code{MedEff_NRT1 }}{plot for hat.mediation1 across t.est}
\item{\code{MedEff_NRT2 }}{plot for hat.mediation2 across t.est}
\item{\code{MedEff_CI_NRT1 }}{plot for hat.mediation1 with CIs across t.est}
\item{\code{MedEff_CI_NRT2 }}{plot for hat.mediation2 with CIs across t.est}
}
}

\examples{
data(smoker)

# GENERATE WIDE FORMATTED MEDIATORS
mediator <- LongToWide(smoker$SubjectID,
                       smoker$timeseq, 
                       smoker$NegMoodLst15min)

# GENERATE WIDE FORMATTED OUTCOMES
outcome <- LongToWide(smoker$SubjectID,
                      smoker$timeseq,
                      smoker$cessFatig)

# GENERATE TWO BINARY TREATMENT VARIABLES
NRT1 <- as.numeric(unique(smoker[,c("SubjectID","varenicline")])[,2])-1
NRT2 <- as.numeric(unique(smoker[,c("SubjectID","comboNRT")])[,2])-1

# GENERATE A VECTOR OF UNIQUE TIME POINTS
t.seq <- sort(unique(smoker$timeseq))

# COMPUTE TIME VARYING MEDIATION ANALYSIS USING BOOTSTRAPPED CONFIDENCE INTERVALS
results <- tvma_3trt(NRT1, NRT2, t.seq, mediator, outcome)

# COMPUTE TIME VARYING MEDIATION ANALYSIS FOR SPECIFIED POINTS IN TIME USING 500 REPLICATES
results <- tvma_3trt(NRT1, NRT2, t.seq, mediator, outcome,
                     t.est = c(0.2, 0.4, 0.6, 0.8),
                     replicates = 500)

}
\references{
\enumerate{
\item{Fan, J. and Gijbels, I. (1996). Local polynomial modelling and its applications: monographs on statistics and applied probability 66 66. CRC Press.}
\item{Fan, J. and Zhang, W. (1999). Statistical estimation in varying coefficient models. The annals of Statistics 27 1491-1518.}
\item{Fan, J. and Zhang, W. (2000). Two-step estimation of functional linear models with applications to longitudinal data. Journal of the Royal Statistical Society: Series B (Statistical Methodology) 62 303-322.}
\item{Cai, X., Piper, M. E., Li, R., & Coffman, D. L. (under review). Estimation and inference for the mediation effect in a time-varying mediation model. Journal of the Royal Statistics Society, Series C (Applied Statistics).}
\item{Baker, T. B., Piper, M. E., Stein, J. H., Smith, S. S., Bolt, D. M., Fraser, D. L., & Fiore, M. C. (2016). Effects of nicotine patch vs varenicline vs combination nicotine replacement therapy on smoking cessation at 26 weeks: A randomized clinical trial. JAMA, 315(4), 371-379. doi:10.1001/jama.2015.19284}
\item{Efron, B.; Tibshirani, R. Bootstrap Methods for Standard Errors, Confidence Intervals, and Other Measures of Statistical Accuracy. Statist. Sci. 1 (1986), no. 1, 54--75. doi:10.1214/ss/1177013815.}
}
}
