% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tvmb.R
\name{tvmb}
\alias{tvmb}
\title{Time-varying mediation function for binary outcome}
\usage{
tvmb(
  treatment,
  t.seq,
  mediator,
  outcome,
  plot = FALSE,
  CI = "boot",
  replicates = 500,
  verbose = FALSE
)
}
\description{
Time-varying mediation function for binary outcome
}
