% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tvmb.R
\name{tvmb}
\alias{tvmb}
\title{Start of the tvmediation function for binary outcome function}
\usage{
tvmb(
  treatment,
  t.seq,
  mediator,
  outcome,
  plot = FALSE,
  CI = "boot",
  replicates = 500,
  verbose = FALSE
)
}
\description{
Start of the tvmediation function for binary outcome function
}
