% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tvmb.R
\name{tvmb}
\alias{tvmb}
\title{Time Varying Mediation Function: Binary Outcome and Two Treatment (Exposure) Groups}
\usage{
tvmb(
  treatment,
  t.seq,
  mediator,
  outcome,
  plot = FALSE,
  CI = "boot",
  replicates = 1000,
  verbose = FALSE
)
}
\arguments{
\item{treatment}{a vector with treatment values}

\item{t.seq}{a vector of unique time points for each observation}

\item{mediator}{matrix of mediator values in wide format}

\item{outcome}{matrix of outcome outcomes in wide format}

\item{plot}{TRUE or FALSE for plotting mediation effect                     Default = "FALSE". (OPTIONAL ARGUMENT)}

\item{CI}{"none" or "boot" method of deriving confidence intervals.       Default = "boot". (OPTIONAL ARGUMENT)}

\item{replicates}{Number of replicates for bootstrapping confidence intervals.    Default = 1000. (OPTIONAL ARGUMENT)}

\item{verbose}{TRUE or FALSE for printing results to screen.                   Default = "FALSE". (OPTIONAL ARGUMENT)}
}
\value{
\item{timeseq}{time points of estimation}

\item{alpha1_hat}{estimated exposure effect on mediator (indirect effect component)}

\item{CI.lower.a1}{lower limit of confidence intervals for alpha1_hat}

\item{CI.upper.a1}{upper limit of confidence intervals for alpha1_hat}

\item{beta1_hat}{estimated exposure effect on outcome (direct effect)}

\item{CI.lower.b1}{lower limit of confidence intervals for beta1_hat}

\item{CI.upper.b1}{upper limit of confidence intervals for beta1_hat}

\item{beta2_hat}{estimated mediator effect on outcome (indirect effect component)}

\item{CI.lower.b2}{lower limit of confidence intervals for beta2_hat}

\item{CI.upper.b2}{upper limit of confidence intervals for beta2_hat}

\item{medEffect}{time varying mediation effect (estimate by product method)}

\item{CI.lower}{lower limit of confidence intervals for medEffect}

\item{CI.upper}{upper limit of confidence intervals for medEffect}
}
\description{
Function to estimate the time-varying mediation effect and bootstrap standard errors, involving two treatment groups and binary outcome.
}
\note{
\enumerate{
\item{Currently supports 2 treatment options. Future releases may expand number of treatment options.}
\item{** IMPORTANT ** An alternate way of formatting the data and calling the function is documented in detail in the function tutorial for tvmb().}
}
}
\section{Plot Returns}{

\enumerate{
\item{\code{plot1_a1 }}{plot for alpha1_hat with CIs across timeseq}
\item{\code{plot2_b1 }}{plot for beta1_hat with CIs across timeseq}
\item{\code{plot3_b2 }}{plot for beta2_hat with CIs across timeseq}
\item{\code{MedEff }}{plot for mediation effects (difference and product) across timeseq}
\item{\code{MedEff_CI }}{plot for CIs of medEffect}
\item{\code{bootstrap }}{plot for estimated medEffects from bootstrapped samples across timeseq}
}
}

\examples{
data(smoker)

# REDUCE DATA SET TO ONLY 2 TREATMENT CONDITIONS (EXCLUDING COMBINATION NRT)
smoker.parsed <- smoker[smoker$treatment != 4, ]

# GENERATE WIDE FORMATTED MEDIATORS
mediator <- LongToWide(smoker.parsed$SubjectID, smoker.parsed$timeseq, smoker.parsed$NegMoodLst15min)

# GENERATE WIDE FORMATTED OUTCOMES
outcome <- LongToWide(smoker.parsed$SubjectID, smoker.parsed$timeseq, smoker.parsed$smoke_status)

# GENERATE A BINARY TREATMENT VARIABLE
trt <- as.numeric(unique(smoker.parsed[ , c("SubjectID","varenicline")])[ , 2])-1

# GENERATE A VECTOR OF UNIQUE TIME POINTS
t.seq <- sort(unique(smoker.parsed$timeseq))

# COMPUTE TIME VARYING MEDIATION ANALYSIS USING BOOTSTRAPPED CONFIDENCE INTERVALS
results <- tvmb(trt, t.seq, mediator, outcome)

}
\references{
\enumerate{
\item{Fan, J. and Gijbels, I. (1996). Local polynomial modelling and its applications: monographs on statistics and applied probability 66 66. CRC Press.}
\item{Fan, J. and Zhang, W. (1999). Statistical estimation in varying coefficient models. The annals of Statistics 27 1491-1518.}
\item{Fan, J. and Zhang, W. (2000). Two-step estimation of functional linear models with applications to longitudinal data. Journal of the Royal Statistical Society: Series B (Statistical Methodology) 62 303-322.}
}
}
