% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tvmcurve_3trt.R
\name{tvmcurve_3trt}
\alias{tvmcurve_3trt}
\title{Main function for time varying mediation function for continuous Outcome and three treatment (Exposure) Groups}
\usage{
tvmcurve_3trt(NRT1, NRT2, t.seq, x, y, t.est)
}
\arguments{
\item{NRT1}{a vector with treatment arm 1 values}

\item{NRT2}{a vector with treatment arm 2 values}

\item{t.seq}{a vector of time points for each observation}

\item{x}{matrix of mediator values in wide format}

\item{y}{matrix of outcome outcomes in wide format}

\item{t.est}{time points to make the estimation             Default = t.seq.}
}
\value{
\item{hat.alpha1}{estimated NRT1 effect on mediator}

\item{hat.alpha2}{estimated NRT2 effect on mediator}

\item{hat.beta3}{estimated mediation effect on outcome}

\item{hat.mediation1}{time varying mediation effect - NRT1 on outcome}

\item{hat.mediation2}{time varying mediation effect - NRT2 on outcome}
}
\description{
Part of the set of internal functions to estimate the time-varying mediation effect and bootstrap standard errors, involving three treatment groups and continuous outcome.
}
