% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tvmcurve_3trt.R
\name{tvmcurve_3trt}
\alias{tvmcurve_3trt}
\title{Main function for time varying mediation function for continuous Outcome and three treatment arms (Exposure Groups)}
\usage{
tvmcurve_3trt(NRT1, NRT2, t.seq, x, y, t.est)
}
\arguments{
\item{NRT1}{a vector with treatment arm 1 values}

\item{NRT2}{a vector with treatment arm 2 values}

\item{t.seq}{a vector of time points for each observation}

\item{x}{matrix of mediator values in wide format}

\item{y}{matrix of outcome outcomes in wide format}

\item{t.est}{time points to make the estimation             Default = t.seq.}
}
\value{
\item{hat.alpha1}{estimated NRT1 (treatment arm 1) effect on mediator}

\item{hat.alpha2}{estimated NRT2 (treatment arm 2) effect on mediator}

\item{hat.gamma1}{estimated NRT1 (treatment arm 1) effect on outcome}

\item{hat.gamma2}{estimated NRT2 (treatment arm 2) effect on outcome}

\item{hat.tao1}{estimated NRT1 (treatment arm 1) effect on outcome, excluding adjustment for mediator}

\item{hat.tao2}{estimated NRT2 (treatment arm 2) effect on outcome, excluding adjustment for mediator}

\item{hat.beta}{estimated mediator effect on outcome}

\item{hat.mediation1}{time varying mediation effect - NRT1 (treatment arm 1) on outcome}

\item{hat.mediation2}{time varying mediation effect - NRT2 (treatment arm 2) on outcome}
}
\description{
Part of the set of internal functions to estimate the time-varying mediation effect and bootstrap standard errors, involving three treatment groups and continuous outcome.
}
