\name{tvmediation}
\title{Time Varying Mediation Analysis}
\alias{tvmediation}

\description{
\if{html}{\figure{tvma-logo.png}{options: align='right'}}

A package for estimating mediation effects that vary over time.
} 

\seealso{
\code{\link{tvma}}, \code{\link{tvma_3trt}}, \code{\link{tvmb}}, \code{\link{LongToWide}}, \code{\link{smoker}}

Useful links:
\itemize{
  \item \url{https://github.com/dcoffman/tvmediation}
}
}

\author{
\strong{Maintainer}: Donna Coffman \email{donna.coffman@temple.edu}

Authors:
	\itemize{
		\item Xizhen Cai
		\item Harry Zobel
		\item Yajnaseni Chakraborti
	}
}
\keyword{internal}