\name{tvmediation}
\title{Time Varying Mediation Analysis}
\alias{tvmediation}

\description{
\if{html}{\figure{tvma-logo.png}{options: align='right'}}

A package for estimating mediation effects that vary over time.
} 

\seealso{
\code{\link{tvma}}, \code{\link{LongToWide}}, \code{\link{smoker}}
}

\author{
Authors:
	\itemize{
		\item Donna Coffman
		\item Xizhen Cai
		\item Harry Zobel
		\item Yajnaseni Chakraborti
	}
}
\keyword{internal}