\name{tvmediation}
\title{Time Varying Mediation Analysis}
\alias{tvmediation}

\description{
\if{html}{\figure{tvma_logo.png}{options: align='right' height = '125'}}

A package for estimating mediation effects that vary over time. This method takes into consideration the dynamic effect that a mediator may have over time; thus the realistic nature of an effect over time is characterized as opposed to the traditional approach of estimating an effect based on its single occurrence.
} 

\seealso{
\code{\link{tvma}}, \code{\link{tvma_3trt}}, \code{\link{tvmb}}, \code{\link{LongToWide}}, \code{\link{smoker}}

Useful links:
\itemize{
  \item \url{https://github.com/dcoffman/tvmediation}
  \item \url{https://github.com/dcoffman/tvmediation/issues}
}
}

\author{
\strong{Maintainer}: Donna Coffman \email{donna.coffman@temple.edu}

Authors:
	\itemize{
		\item Xizhen Cai
		\item Harry Zobel
		\item Yajnaseni Chakraborti
	}
}
\keyword{internal}